requirements.txt
appdirs==1.4.1
numpy==1.12.0
packaging==16.8
PyOpenGL==3.1.0
pyparsing==2.1.10
six==1.10.0
